+++ Sources/CrabCore/CrabCore.swift
@@ -1,3 +1,3 @@
 public enum CrabCore {
-    public static let version = "0.2.5"
+    public static let version = "0.2.6"
 }